# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  version: 1.5.0
  description: API для создания, чтения, поиска, обновления задач и управления тегом задачи.
servers:
  - url: /
tags:
  - name: Tasks
    description: Операции с задачами
paths:
  /v1/tasks:
    get:
      tags:
        - Tasks
      summary: Найти задачи
      description: Возвращает список задач. Сейчас поддерживается фильтр только по `title`.
      operationId: queryTasks
      parameters:
        - in: query
          name: title
          required: false
          description: Фильтр по заголовку задачи
          schema:
            type: string
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      responses:
        '200':
          description: Список задач
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadTask'
              examples:
                found:
                  value:
                    - id: 42
                      title: Buy milk
                      status: open
                      description: ""
                      category: ""
                      tags: []
                      created_at: "2026-05-20T12:00:00Z"
                    - id: 43
                      title: Buy bread
                      status: done
                      description: ""
                      category: home
                      tags:
                        - grocery
                      created_at: "2026-05-20T13:00:00Z"
                empty:
                  value: []
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
    post:
      tags:
        - Tasks
      summary: Создать задачу
      description: Создаёт новую задачу и возвращает сохранённую сущность.
      operationId: createTask
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
            examples:
              minimal:
                summary: Минимальный запрос
                value:
                  title: Buy milk
                  status: open
              full:
                summary: Полный запрос
                value:
                  title: Prepare sprint board
                  status: in_progress
                  description: Review backlog and create tasks
                  category: work
                  tags:
                    - planning
                    - sprint
      responses:
        '201':
          description: Задача создана
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadTask'
              examples:
                created:
                  value:
                    id: 42
                    title: Buy milk
                    status: open
                    description: ""
                    category: ""
                    tags: []
                    created_at: "2026-05-17T12:00:00Z"
        '400':
          description: Некорректный JSON или ошибка валидации
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidJson:
                  summary: Некорректный JSON
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid json body
                validationError:
                  summary: Ошибка валидации
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: validation_error
                    message: validation error
                invalidStatus:
                  summary: Невалидный статус
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid status
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
  /v1/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Получить задачу
      description: Возвращает задачу по её идентификатору.
      operationId: readTask
      parameters:
        - in: path
          name: id
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      responses:
        '200':
          description: Задача найдена
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadTask'
              examples:
                found:
                  value:
                    id: 42
                    title: Buy milk
                    status: open
                    description: ""
                    category: ""
                    tags: []
                    created_at: "2026-05-17T12:00:00Z"
        '400':
          description: Некорректный идентификатор задачи
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidTaskID:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: validation_error
                    message: invalid task_id
                emptyTaskID:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: validation_error
                    message: task_id is empty
        '404':
          description: Задача не найдена
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                taskNotFound:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 404
                    code: resource_not_found
                    message: task not found
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
    patch:
      tags:
        - Tasks
      summary: Обновить задачу
      description: Частично обновляет задачу по её идентификатору.
      operationId: updateTask
      parameters:
        - in: path
          name: id
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
            examples:
              title:
                summary: Обновление заголовка
                value:
                  title: Updated Title
              full:
                summary: Обновление нескольких полей
                value:
                  title: Prepare release
                  status: in_progress
                  description: Finalize checklist
                  category: work
      responses:
        '200':
          description: Задача обновлена
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadTask'
              examples:
                updated:
                  value:
                    id: 42
                    title: Prepare release
                    status: in_progress
                    description: Finalize checklist
                    category: work
                    tags: []
                    created_at: "2026-05-19T12:00:00Z"
        '400':
          description: Некорректный JSON, ошибка валидации или неверный статус
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidJson:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: validation_error
                    message: unexpected EOF
                invalidTaskID:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: validation_error
                    message: invalid task_id
                invalidStatus:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid status
                noChanges:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: no changes in update task request
        '404':
          description: Задача не найдена
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                taskNotFound:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 404
                    code: resource_not_found
                    message: task not found
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
  /v1/tasks/{id}/tags:
    post:
      tags:
        - Tasks
      summary: Добавить тег к задаче
      description: Добавляет один тег к задаче по её идентификатору.
      operationId: addTaskTag
      parameters:
        - in: path
          name: id
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
            examples:
              add:
                value:
                  tag: urgent
      responses:
        '200':
          description: Тег добавлен
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadTask'
        '400':
          description: Некорректный идентификатор, JSON или тег
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidID:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid id
                invalidJson:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid json body
                invalidTag:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: tag is required
        '404':
          description: Задача не найдена
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                taskNotFound:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 404
                    code: resource_not_found
                    message: task not found
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
    delete:
      tags:
        - Tasks
      summary: Удалить тег у задачи
      description: Удаляет один тег у задачи по её идентификатору.
      operationId: deleteTaskTag
      parameters:
        - in: path
          name: id
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: header
          name: X-Request-ID
          required: false
          description: Пользовательский request ID. Если не передан, сервер сгенерирует его сам.
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
            examples:
              delete:
                value:
                  tag: urgent
      responses:
        '200':
          description: Тег удалён
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadTask'
        '400':
          description: Некорректный идентификатор, JSON или тег
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidID:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid id
                invalidJson:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: invalid json body
                invalidTag:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 400
                    code: bad_request
                    message: tag is required
        '404':
          description: Задача или тег не найдены
          headers:
            X-Request-ID:
              description: Идентификатор запроса
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                taskNotFound:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 404
                    code: resource_not_found
                    message: task not found
                tagNotFound:
                  value:
                    request_id: clx7f4b2f0000qtr8gk1d0abc
                    status_code: 404
                    code: resource_not_found
                    message: tag or task not found
        '500':
          description: Внутренняя ошибка сервера
          headers:
            X-Request-ID:
              description: Идентификатор запроса может присутствовать в ответе
              schema:
                type: string
components:
  schemas:
    TaskStatus:
      type: string
      description: Статус задачи
      enum:
        - open
        - in_progress
        - done
        - closed
    CreateTask:
      type: object
      additionalProperties: false
      required:
        - title
        - status
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Заголовок задачи
        status:
          $ref: '#/components/schemas/TaskStatus'
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: Описание задачи
        category:
          type: string
          minLength: 1
          maxLength: 500
          description: Категория задачи
        tags:
          type: array
          minItems: 1
          maxItems: 10
          description: Список тегов
          items:
            type: string
            minLength: 1
            maxLength: 30
    UpdateTask:
      type: object
      description: Частичное обновление задачи. Нужно передать хотя бы одно поле.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Новый заголовок задачи
        status:
          $ref: '#/components/schemas/TaskStatus'
        description:
          type: string
          maxLength: 500
          description: Новое описание задачи
        category:
          type: string
          maxLength: 500
          description: Новая категория задачи
    Tag:
      type: object
      required:
        - tag
      properties:
        tag:
          type: string
          minLength: 1
          maxLength: 30
          description: Тег для добавления или удаления
    ReadTask:
      type: object
      required:
        - id
        - title
        - status
        - description
        - category
        - tags
        - created_at
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - request_id
        - status_code
        - code
        - message
      properties:
        request_id:
          type: string
        status_code:
          type: integer
          format: int32
        code:
          type: string
          enum:
            - bad_request
            - validation_error
            - resource_not_found
        message:
          type: string